requirements.txt
python-telegram-bot==22.1
openai==2.29.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2